test: update cases
